+++ src/main/java/com/example/production/model/RawMaterial.java
@@ -18,7 +18,9 @@ public class RawMaterial {
     @ManyToOne(cascade = CascadeType.MERGE)
     @JoinColumn(name = "unit_of_measurement_id", referencedColumnName = "id")
     private UnitsOfMeasurement unitsOfMeasurement;
-    private double quantity;
-    private double amount;
+//    @Convert(converter = DoubleToStringConverter.class)
+    private float quantity;
+//    @Convert(converter = DoubleToStringConverter.class)
+    private float amount;
 }
